Pass build directory from builder to tester.
In switching builders to set the swarming command line directly,
we had been relying on the tester having the same build directory
configured as the builder (e.g., "out/Debug").

However, if you have a 64-bit builder running on Windows and a
thin tester running on Linux, by default you'll build into
"out/Release_x64" and try to use "out/Release" on the tester.

We should've just been passing the build directory to use along
with the other information on the command line, so as to not rely
on this. This CL adds an additional build property,
'swarming_command_lines_cwd', to do that.

Bug: 1116139, 1108005
Change-Id: I68d29532fa2acebfe12f5c1c27930ffe00926fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2354631
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>