Flip chromium.mac ios bots to send swarming command lines.
This changes the public chromium.mac iOS bots to include the swarming
command lines directly in the tasks, rather than relying
on the bots to pull the command lines from the isolates.

Apparently I was confused and thought they weren't using the
chromium recipe and didn't need to be changed, or something.

Bug: 1108005

Change-Id: I2ed35d5a96a8b4734b4bfe9c1a60d4bde9674168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2357495
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Jeff Yoon <jeffyoon@chromium.org>