Add missing shards to the list of failed shards
If a test (collection of swarming tasks, in our case a test suite)
marks a shard as missing, we currently only change the UI of the build,
and mark the test suite itself as invalid, but don't do anything else.
This unfortunately breaks the 'retry shards with patch' logic. That
logic expects tests to list the failed shards in it that should be
retried. The logic which currently sets that list only looks at the
swarming summary.json file, which means that if a merge script changes
the test results, but not the swarming summary json file, then the shard
will not be marked as a failed shard, and the recipe will fail because
it's trying to retry a test suite which is marked as invalid, but
doesn't appear to have any failed shards.

This CL changes the behavior to add the list of missing shards in the
test results json to the list of failed shards stored in the task
object. See the bug for more details.

Bug: 1119972
Change-Id: I4184739f6cd45dfb93a00f80f76bb3fba453c0ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2368052
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Auto-Submit: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>