Skip analyze when one of the test spec files is affected.
Detecting if any of the spec files that we've used are affected by the
CL allows us to skip performing analysis and just run all of the tests
for the builder. This will allow us to exit early from builders that
don't consumed the affected spec files for CLs that only affect the spec
files.

Bug: 1123187
Change-Id: I2cafafa83370d9455e705d29b2ee0ef0cc0359b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2385922
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>