Add a presubmit to prevent using deprecated properties methods.
The properties test API methods set properties as they would be set for
a buildbot build, not as they would be set for a LUCI build. One of the
properties that they set is the mastername property, so new test cases
have been added that prevent the removal of the mastername property. The
presubmit will prevent the addition of new uses of the deprecated
properties methods which add the mastername property in a
non-forward-compatible way.

Bug: 1109276
Change-Id: I7fc80e0a2f0760897239b651c30f14c77496c73a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2389083
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>