Refactor chromium_test generator functions.
In order to simplify the implementation needed to fix crbug.com/1131231,
this CL does some minor refactoring of the generator functions in
the generators module of the chromium_tests recipe_module, in order
to allow the gtest generator to be able call the isolate generator
directly.

No functional changes are expected.

Bug: 1131231
Change-Id: I7e47bc4f3a6e9e42e68beef1bfb5fbc433a2dc83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2425209
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>