Remove --isolated-script-test-perf-output from chromium recipe.
As discovered/discussed in crbug.com/1127545, it appears that this
flag is not actually ever used anywhere in the chromium recipes.

Perf tests *do* produce output, but they don't use this flag to
do so.

WebRTC tests *do* use this flag, but they set this directly in their
own recipes.

Bug: 1127545
Change-Id: I215ad8f698c18e41ca2a87c7043371a9ac73000b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2427305
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>