Add use_isolated_scripts_api key to gtest_tests chromium_test generator.
In order to migrate the non-isolated-script-based tests to the
isolated-scripts API, we either have to move tests around
in test_suites.pyl, or make it possible to separate what a test
is from what group it is part of.

This CL takes the latter approach: it adds a new key,
'use_isolated_scripts_api' to the 'gtest_tests' test_spec declared
in the //testing/buildbot/*.json files. If that key is set to True,
we will treat that test as if it was declared in 'isolated_scripts'
instead of in 'gtest_tests' (i.e., create a SwarmingIsolatedScriptTest
for it instead of a SwarmingGtestTest).

Bug: 1131231
Change-Id: I0cefd1eb1d4d97f9e54626a1545e7f67d0a59482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2425210
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>