Rework analyze DEPS logic to run before testing begins
Previously, this logic ran at the end of a build. This ran into the problem of the patch being deapplied in the "without changes" phase and thus this logic couldn't see which files were changed. Now it's split up to get a revised list of test targets during the analyze phase, while still doing the correctness checking at the end.

Bug: 923016
Change-Id: I120ff9d463c05d02515ed81ca83005ad9cb42ca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2424888
Commit-Queue: Gregory Guterman <guterman@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>