Make the path to merge scripts settable.
The path to testing dir was only being used for getting the path to the
merge scripts, so reducing the scope of the configuration to that makes
it clearer what would need to be duplicated to have a different merge
scripts directory. The code for getting a merge script was modified to
fall back to using the chromium_checkout working directory so that it
wouldn't be necessary to call configure_swarming.

Change-Id: I223b2ddad7c9fab47dabe0f8d11160401ea41d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2485894
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>