Make the iOS recipe module pass 'command' to Swarming.
This CL adds the possibility to pass 'command' to swarming at trigger
time instead of assuming it will be read from the .isolated file.

The 'command' and 'relative_cwd' fields are deprecated but they will
still be in the .isolated file so this CL reads them from it and
keep them in memory until the tests are triggered.

In the future, if 'command' and 'relative_cwd' will be removed from the
.isolated file, this idea will stop working. Since WebRTC is the
only used of this recipe_module, WebRTC should move away from it
and switch to the canonical way of getting the 'command line' via
MB and GN.

Led runs:

* WebRTC ios-sim: https://chromium-swarm.appspot.com/task?id=4ec890be775d4510
* WebRTC/Chromium FYI: https://chromium-swarm.appspot.com/task?id=4ec891068072a310

Recipe-Nontrivial-Roll: build_limited_scripts_slave
Bug: webrtc:11604
Change-Id: Id8a6277ce5e82f6832866afc2108b632ee9d247e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2418404
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>