Make failing to find the isolate for a swarming test infra fail.
Failure to find the isolate for a swarming test will rarely indicate a
problem with the chrome binaries, instead it will almost always
indicate a problem with the LUCI services config or the recipe. Due to
that, it's better to have the failure to find the isolate result in an
infra failure so that it will generate alerts and the infra team can
look into it rather than waiting until we get notified about it.

Change-Id: I61a0f3e1bca86ac41757a696f089a552524449cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2490236
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>