Make all local script tests have invalid results on non-zero exit code.
This will make any local-script test that has no failures reported in
its json still be a failure if it exits with a non-zero code.

eg: This build is green despite webkit_lint exiting with a non-zero code
https://ci.chromium.org/p/chromium/builders/try/linux-rel/518410?

With this patch, that build no becomes red. eg:
https://ci.chromium.org/swarming/task/4f4c52cbd3076910

Bug: 1108689
Change-Id: I4649af21c587e19869d0ec9c0753fae7a136f006
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2476959
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>