[ios/webrtc] build command and pass it to Swarming
command and relative_cwd properties are being deprecated from
isolate and isolated (crbug/1069704). A patch was introduced
in crrev/c/2418404 to pass command by fetching what was set in
isolated. However, isolated will no longer have the command,
so this results in a KeyError.

This change builds the command and appends it to the task dict
in the isolate_test() method as a workaround with raw_cmd as key.
relative_cwd remains defaulted to ".".

The SwarmingIosTest object takes raw_cmd and relative_cwd from
this dict and passes them when creating the Swarming task object
in pre_run.

Bug: 1144638
Change-Id: I969e81779880b3b2057f041f427b4614c17fd161
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2518615
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>