Expose mirrored builders as an output property
Exposes the list of builders that mirror a particular CI builder as an
output property, which is a list of strings. Strings are in a
"group:builder" format.

This is intended to be used to be able to programatically determine
which trybots need to be queried for ResultDB data given some list of
CI builders, but could also be potentially useful to users by allowing
them to see which trybots mirror a particular builder without having to
look at the trybot mapping.

Bug: 1145293
Change-Id: I33153b81929910cc11586382770cec63163b032f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2518663
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>