Remove **runtest_kwargs from Test subclass __init__ methods.
Taking in arbitrary keyword arguments to be passed to runtest is mostly
unused: the generator for local gtests will never set any arbitrary
keyword arguments and WebRTCPerfTest sets it, but needs to make
modifications to it as part of running the test because it needs to read
values from the API to determine the final value. The classes have been
modified with a _get_runtests_kwargs method that returns {} for
LocalGTestTest and returns a dictionary to configure the perf uploading
for WebRTCPerfTest.

In the case of LocalIsolatedScriptTest, runtest_kwargs is not used by
the class, so taking arbitrary keyword arguments in runtest_kwargs is
causing keyword ResultDB-related keyword arguments provided by the
generator (full_test_target, test_id_prefix) to be effectively ignored.
Named arguments have been added so that they can take effect as
expected.

This simplifies further refactoring to remove the test_specs field from
bot_spec.BotSpec.

Bug: 1128746
Change-Id: I168e93faa3a3124850de0ed6ef2942e260f5ac86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2514903
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>