Convert None to the default for attribs.
This simplifies code that initializes attrs-based objects from
dictionaries that may or may not have the corresponding keys. Instead
of having to check if attribute 'x' is  present in the dict to determine
whether or not to pass an argument for the attribute, you can simply
pass x=d.get('x').

Bug: 1128746
Change-Id: I17a6147fdb9599ba7d1d8b5a3a8e10b045105a62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2513752
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>