Modify required attributes to not need to appear before optional.
This will enable the creation of test spec types for each test type that
accurately indicate required fields. For example ScriptTest requires
that the script value be provided. This wouldn't be possible to declare
using the standard behavior no default behavior because it would appear
on a spec type that inherits from the base spec type which will have
optional attributes. That would require a workaround of accepting a None
default and customizing the create method to require the parameter.

Bug: 1128746
Change-Id: I41ce870d695a83451312db60ff68183e1e233e48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2518000
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>