tricium: Add a recipe wrapper for simple chromium analyzers
Context: Because we want to remove dependency on isolate server api
(and move to a recipe-based-analyzer-only model for the future
LUCI Change Verifier), we want to run all existing analyzers in
recipes.

This adds an initial recipe that runs existing non-recipe-based
analyzers for the chromium project.

Bug: 1105239
Change-Id: I373a1e717a27c8467f722392f2344c345195a1b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2477246
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>