Configure the chromium_android module for all android builders.
Curently, if an android builder does not specify an android config, it
ends up causing an infra failure when it attempts to run
chromium_android module code because the module has not been configured.
With this change, the same condition gates whether to configure the
chromium_android module and whether to call its code. It is not
necessary to explicitly set the android config in a bot spec because the
module has a root config which sets necessary values. The majority of
configs for the module do not have any effect and should be cleaned up.

Change-Id: I8afb7278173a253abc2bbeeeeee496a88f0fd74b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2525788
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>