[weblayer, skew tests] Add CIPD metadata to step_metadata
We plan on using a floating reference for downloading CIPD packages
in skew tests configurations. By using a floating reference we can
refresh skew tests without having to land CL's in src/. We would just
change the CIPD package which the floating reference points to in
the CIPD server. This change will expose information about the CIPD
packages for a test suite. If we add a tag to the CIPD package
specifying which Chromium version it contains, then users
can look in the test suite metadata and find out which version is being
skew tested by the builder.

Bug: 1147459, 1147455, 1099017, 1041619
Change-Id: If0ca9d6e94aae2c05225892a280658a2948f279b
Recipe-Nontrivial-Roll: chrome_release
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2533379
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Michael Moss <mmoss@chromium.org>