[resultdb] add ResultDB.wrap
chromium_swarming/api implements the following functions to wrap
a given cmd with result_adapter and ResulSink.
- _wrap_cmd_with_result_adapter()
- _maybe_enable_resultdb_for_task()

Similarly, chromium_android and build.ToolsBuildApi need
those functions to support ResultDB on LocalGTestTest and
AndroidJUnitTest.

This CL adds wrap() to steps.ResultDB so that each of the test runner
modules don't need to implement it, but can use the given ResultDB
instance to wrap the command as necessary.

Bug: 1084332
Change-Id: I3387506813a61460cc5143c1e4ddd89557e47392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2574662
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>