tricium_clang_tidy: emit multiple diags per use-after-move
With bugprone-use-after-move, it's possible to introduce a diagnostic on
line N from line M. Because of this && because we filter message
emission based on line number, it's probably best to emit these at each
place that tidy notes something bad is happening.

This also removes an import that's been useless for a while.

Bug: 1157503
Change-Id: I18ad9c29318002e209e1723cc0b95a75d4bb4f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2586258
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>