[js-code-coverage] Add get_files_coverage_data helper method
Introduce a method that starts to generate the end format required to
integrate into the code coverage dashboard. This extracts all the source
files from the coverage files, retrieves the covered lines, uncovered
lines, relative source path and adds a scaffolding method which returns
the file summary level metrics.

The metrics for now are only line coverage and there is an explicit
Exception raised if duplicate source files are found in >1 coverage
file. There is no possiblity for now for this to occur hence raising an
Exception.

Bug: 1113941
Test: ./recipes.py test run
Test: ./generate_coverage_metadata_for_javascript_test.py
Test: python generate_coverage_metadata_for_javascript.py \
	--src-path chromium/src \
	--coverage-dir chromium/src/out/Default/devtools_code_coverage \
	--dir-metadata-path component_mapping \
	--output-dir none
Test: https://chromium-swarm.appspot.com/task?id=5077e6e79e10c310
Change-Id: I8c70637460a9922718ac65bf13818f3613c19cab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2588506
Commit-Queue: Ben Reich <benreich@chromium.org>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>