Adding support for chain-of-custody signatures for upload files
Files for upload can add chain-of-custody signatures using the
'verifiable_key_path' key. This string path will invoke cloudkms
to create a .sig file for each file in the ArchiveData definition.
The .sig file is to be uploaded next to the original file in gcs.
This .sig file will also be bundled into the .zip or .tar.gz if
that's the selected archive type.

Change-Id: If42674653894885bfa67cce665795044c51a6cfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2586006
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>