Add trace tests to ANGLE CI.
This is an alternate testing mode that does all the ninja
and GN work in the python script. The trace bot is only run
manually and in CI jobs. Testing of this change via 'led'.
The trace tests turning red won't block the CQ.

Bug: angleproject:5530
Change-Id: I1370877682654c384e10248cf99e9512dd0cd7b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2639311
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Jamie Madill <jmadill@chromium.org>