Trigger child builders using a GitilesTrigger.
The GitilesTrigger allows for specifying the gitiles commit. This
enables propagating the ref information so that the gitiles information
of the triggered builder can match that of the triggering builder
without having to write ad-hoc code.

This will allow the Android official testers to set the output commit.
The Android official testers are run on tags, which are applied to
commits that publish DEPS for the version and do not contain a commit
position, which causes the set output commit feature to fail because
either a ref must be provided or the checked out revision must have a
commit position footer.

Bug: 1141886
Change-Id: I39aa7db590a8fbf31b17c9cbe88431d1eb033590
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: chrome_release
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2668040
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>