Remove normalize methods from the chromium_tests types.
This will force config objects to be initialized with the correct type,
simplifying code by allowing it to expect a single type. TryMirror
retains the ability to normalize a BuilderId because it is a nested type
that is not often passed around on its own and an upcoming refactoring
should eliminate TryMirror to instead just use a collection of
BuilderId.

Change-Id: Ia16362e478a62a76f11c66dbfe44cb390d0991d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2679107
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Michael Moss <mmoss@chromium.org>