Use tast_expr_b64 to pass base64 encoded tast expression
Due to crbug/1173329, skylab does not support arbitrary tast
expressions now. The fix may not land in the foreseen future. As a
workaround we encode the expression in chromium recipe and decode it
in autotest(done in crrev.com/c/2666384), so that any legal tast
expressions can travel through skylab.

BUG=chromium:1139516
TEST=train

Change-Id: Ie1bc35a4cd5d3547cd5ab0575a66c44338ac2c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2674827
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Xinan Lin <linxinan@chromium.org>