boringssl: Make it possible to specify Windows ARM64 builder
This, along the way, changes up how we configure the recipe a bit.
Historically, we parsed the options out of the builder name. That meant
adding an entry to CI required adding a new name option to the recipe,
waiting for that to roll out, and then adding the new builder to
infra/config.

Instead, the preferred mechanism for configuring a recipe is properties.
I've added a pile of generic properties, which should be sufficient to
configure a Windows ARM64 compile-only bot. Hopefully this also means
we'll need to touch fewer repositories to add a builder next time.

This makes many of the name-based options redundant, but the existing
builders don't know to specify them yet. As a follow-up, I'll switch the
infra/config to specify the same options explicitly, which should let us
trim this recipe down significantly. I'll probably need some extra
iterations beyond that, but this should be a good start. In particular,
it is hard to specify properties that include the checkout path.

The tests should demonstrate that nothing changes for the existing
name-based scheme. Note the 'compile' bots *do* change slightly; they
now always run check_filenames.go, which is harmless. It seemed tidier
to have separate run_unit_tests and run_ssl_tests properties.

(This framework is a bit annoyingly magical and tries to explode
properties into the parameters of RunSteps, which makes it hard to pass
them into helper functions. I've turned that around and wrapped them
back into a Config object. This is a little silly.)

Bug: boringssl:397
Change-Id: I0e42db90a5c596e71c1fd8b90caaee666fee209d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2692331
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>