[Chromium recipes][ResultDB] Remove the limit on chromium ci/try
If the build has an invocation, we should try to query invocations and test results for it.

As a result, we can display test results on milo for chromium branch builds, if resultdb is enabled on that build.

Bug: 1179357
Change-Id: I0caad769fcd543103f4531ed141b3ff5b6b6a8b5
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: chrome_release
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2716093
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>