[V8] Infer if CAS was used when building
The / in the digest differentiates the IDs between CAS and isolate.
We now use this information when triggering tests, so that the method
always fits to what the builders have used.

This is less disruptive when flipping builder/tester pairs, since now
the ongoing builds don't need to be switched in sync.

Bug: 1154223
Change-Id: I9cd8105da6897b7a29677d5c7a4359cea84f1c2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2742493
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Michael Moss <mmoss@chromium.org>