[js-code-coverage] Extend JS coverage module to work per-cl
In CL:2721892 and CL:2661439 a new optional trybot was added that can be
opted into. Update the code_coverage recipe module to not bail early
when it's identified as per-cl mode and extend it to take parameters
pertaining to per-cl coverage.

JavaScript coverage can be ran at the same time as other languages so to
ensure the coverage is separate we track the affected JS source files in
another list and kick off coverage processing only when that list has
values.

Appropriate TODO's have been left for follow up work such as actually
utilising the diff mapping and source files.

Bug: 1113941
Change-Id: Ic3b80436e4b7a3bcf6c5e7738acc26abf1b44e82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2728583
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>