goma_client recipe: Add some support for a bot that sets arch to arm
With this setup, we'll build a dedicated cipd package for mac-arm64.
It assumes that we'll teach build/archive.py in the client.git repo
to accept `--platform=mac-arm64` for mac/arm64 builds.

It also requires that we update cipd_manifests.txt in depot_tools to
pull goma/client from ${platform} instead of from ${os}-amd64 like
currently.

And we'll have to create new bots that set platform.arch to 'arm',
somehow, in
https://chrome-internal.googlesource.com/infra/infra_internal.git/+/refs/heads/main/infra/config/subprojects/goma.star

(And, if you want try converage for this, also in several of the files in
https://chromium.googlesource.com/infra/goma/client/+/refs/heads/infra/config )

Bug: b/183118231
Change-Id: I0c619e5ba034f9b36f9579d21e31961baeac8160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2780615
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>