Add ability to only run tests post-submit in chromium recipe.
Sometimes it would be nice to be able to configure a test to
only run post-submit (on the continuous builders), without also
requiring it to run pre-submit on the tryjob/CQ builders.

This CL adds the ability to look for a new boolean flag
called `ci_only` in the test spec; if set to true, we will only
run the test if the bot is not a tryserver.

Bug: 1170989
Change-Id: Ic476a1d617a54c3369c2f2c58d2b0bd15c9f2054
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2783046
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>