Add constraints to attr_utils.
Constraints will provide the ability to provide deeper and more flexible
type-checking of fields on classes defined using attrs. It will no
longer be necessary to specify a specific sequence or mapping type when
they appear as the member type of a sequence or the key or value type of
a mapping: sequence or mapping can be used as appropriate. The
constraints can also be combined and nested, so for example one could
enforce the keys and values of mappings that appear in a sequence have
certain types.

Change-Id: I628b9701f2051b12efb1a1b67a0aece8eeda08a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2792049
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>