Use the builder spec when looking up bulder GN args.
When looking up the builder GN args, there is at most one builder
involved and the operation relies only on information that is present in
the spec for that builder, so there is no need to create a bot config
object. Additionally, the _chromium_config method is only used in that
single spot and it's a small enough block of code to contain it directly
at the callsite.

Change-Id: I0da41c91b67cf4b0d7c244d5f67672ee5f9eb919
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2795281
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>