Restore symupload module to build repo.
When the builders are moved to the bootstrapping model where there is no
configuration for the builder specified within the recipe, there will no
longer be a need to have separate internal recipes to hold the
configuration for internal builders if all of the functionality is
contained in the pulic recipe. The symupload functionality is the only
major piece of functionality currently present in the internal recipe
that is not present in the public recipe.

The module is up to date with the most recent version of the symupload
module from the downstream repo.

Bug: 1192714
Change-Id: I54fea5f2587153050f706911e510fcf3421d77a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2787245
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Michael Moss <mmoss@chromium.org>