angle: Add source-side GN configuration.
This uses the "mb" path to configure the ANGLE builders. It introduces
an ANGLE recipe module for use with the chromium_trybot and other
modules. Also includes several configs that we'll use to call into
the helper modules. In the future it will lead to uploading and
triggering tests directly from the ANGLE recipe.

Bug: angleproject:5114
Change-Id: Iad0c527794e23aff6809e4c230f1d8ace92bd1de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2822692
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Jamie Madill <jmadill@chromium.org>