Upgrade the CELab Go version to latest (1.16.3)
CELab has been using Go version 1.11.2, which is several years old at
this point, and severely out of sync with a Google engineer's
development environment.

This change updates the version of Go used for CELab buildbots to bring
it in sync with the development environment. Following this change will
be a series of CLs that addresses build breaks that comes with the
upgrade.

Bug: 1199031
Change-Id: I05c2761588d4e890c5436571ca95797f5815d696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2842676
Commit-Queue: Yiming Zhou <uwyiming@google.com>
Reviewed-by: Xiang Ji <jxiang@google.com>