Remove the path-based iOS testing.
The paths that the iOS testing was being triggered on haven't existed
for some time, since the repo was updated to move the root of the
recipes. Additionally, the ios module and recipes have been removed with
the iOS builders having been switched to use the same recipes as the
other OSes.

Change-Id: I92c04297fb1de3c203e717bb07abd3872c912f95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2847469
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>