Simplify the configuration of builders to trigger.
This puts all the details of the builders to trigger in one spot, making
it easier to update the builders. It also removes the need to update the
list of builders for branches because branched versions of the default
builders will now be looked up for all of the buckets that we have set
CLs for.

Change-Id: Ia9a7567a4d3ca0c4040e7aaba327933065a635ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2847263
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>