Group all RDB invocation results by variant hash throughout test_utils.
And use it to inform the "which suites to retry" decision logic for the
new RDB-flow.

As-is, the current logic counts any shard with an unexpected result
(including an unexpected PASS) as a retriable suite. eg: See the
diff in retriable suite count in
https://ci.chromium.org/old/p/chromium/builders/try/linux-rel/675376
ie:
Legacy decision for retries had length 2
New decision for retries had length 39

By counting suites instead of shards and ignoring unexpected PASSes,
that discrepancy should narrow down significantly.

Note that we need to incorporate suites with shards that never ran
(eg: EXPIRED, NO_RESOURCE) in the decision logic for the new flow. But
since RDB isn't aware of those types of results (which is WAI), we need
to maintain some awareness of the results of the shards. This is done
by calling has_valid_results() in the new flow, with the goal to strip
down has_valid_results() to remove json-parsing logic in future CLs.

LED build with this patch applied & same test failures as the linux-rel
build above:
https://ci.chromium.org/swarming/task/5328827fbd970010?server=chromium-swarm.appspot.com

Bug: 1135718
Change-Id: I00a5c3173bec531750d944dd316bfcdc06878648
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2847030
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>