Refactor results returned by RDB into its own class.
This tears out the elaborate dict that we were using to store RDB's
results, and moves them into a class defined in util.py. Any
subsequent inspection of the results will instead be handled by
accessing helper methods/fields of the class, rather than iterating over
the elaborate dict every time.

This also fixes a small bug where the new RDB decision logic would
falsely consider a test with different types of failures as an
unexpected failure. Specifically, the old logic would consider a
test that's expected to CRASH and that runs with [ABORT, CRASH] as
unexpected. But it's actually expected since it finishes with CRASH.

Bug: 1135718
Change-Id: I6dbeb20d2bf82f10fcc6d678de7b2b35eb4f95a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2857696
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>